+++ pyhap/accessory_driver.py
@@ -57,12 +57,12 @@ class AccessoryMDNSServiceInfo(ServiceInfo):
     def __init__(self, accessory, address, port):
         self.accessory = accessory
         hname = socket.gethostname()
-        pubname = hname + "." if hname.endswith(".local") else hname + ".local."
+        pubname = hname + '.' if hname.endswith('.local') else hname + '.local.'
 
         adv_data = self._get_advert_data()
-        super(AccessoryMDNSServiceInfo, self).__init__(
-             "_hap._tcp.local.",
-             self.accessory.display_name + "._hap._tcp.local.",
+        super().__init__(
+             '_hap._tcp.local.',
+             self.accessory.display_name + '._hap._tcp.local.',
              socket.inet_aton(address),
              port,
              0,
@@ -78,26 +78,24 @@ class AccessoryMDNSServiceInfo(ServiceInfo):
 
     def _get_advert_data(self):
         """Generate advertisement data from the accessory."""
-        adv_data = {
-            "md": self.accessory.display_name,
-            "pv": "1.0",
-            "id": self.accessory.mac,
-            # represents the "configuration version" of an Accessory.
-            # Increasing this "version number" signals iOS devices to
+        return {
+            'md': self.accessory.display_name,
+            'pv': '1.0',
+            'id': self.accessory.mac,
+            # represents the 'configuration version' of an Accessory.
+            # Increasing this 'version number' signals iOS devices to
             # re-fetch accessories data.
-            "c#": str(self.accessory.config_version),
-            "s#": "1",  # "accessory state"
-            "ff": "0",
-            "ci": str(self.accessory.category),
-            # "sf == 1" means "discoverable by HomeKit iOS clients"
-            "sf": "0" if self.accessory.paired else "1",
-            "sh": self._setup_hash()
+            'c#': str(self.accessory.config_version),
+            's#': '1',  # 'accessory state'
+            'ff': '0',
+            'ci': str(self.accessory.category),
+            # 'sf == 1' means "discoverable by HomeKit iOS clients"
+            'sf': '0' if self.accessory.paired else '1',
+            'sh': self._setup_hash()
         }
 
-        return adv_data
 
-
-class AccessoryDriver(object):
+class AccessoryDriver:
     """
     An AccessoryDriver mediates between incoming requests from the HAPServer and
     the Accessory.
@@ -108,8 +106,8 @@ class AccessoryDriver(object):
 
     NUM_EVENTS_BEFORE_STATS = 100
 
-    def __init__(self, accessory, port, address=None, persist_file="accessory.state",
-                 encoder=None):
+    def __init__(self, accessory, port, address=None,
+                 persist_file='accessory.state', encoder=None):
         """
         :param accessory: The `Accessory` to be managed by this driver. The `Accessory`
             must have the standalone AID (`pyhap.accessory.STANDALONE_AID`). If the
@@ -157,15 +155,15 @@ class AccessoryDriver(object):
             self.persist()
         self.topics = {}  # topic: set of (address, port) of subscribed clients
         self.topic_lock = threading.Lock()  # for exclusive access to the topics
-        self.event_loop = asyncio.new_event_loop()
-        self.aio_stop_event = asyncio.Event(loop=self.event_loop)
+        self.loop = asyncio.new_event_loop()
+        self.aio_stop_event = asyncio.Event(loop=self.loop)
         self.stop_event = threading.Event()
         self.event_queue = queue.Queue()  # (topic, bytes)
         self.send_event_thread = None  # the event dispatch thread
         self.sent_events = 0
         self.accumulated_qsize = 0
 
-        self.accessory.set_broker(self)
+        self.accessory.set_driver(self)
         self.mdns_service_info = None
         self.srp_verifier = None
         self.accessory_thread = None
@@ -231,7 +229,7 @@ class AccessoryDriver(object):
         this is not run in a daemon thread or it is run on the main thread, the app will
         hang.
         """
-        while not self.event_loop.is_closed():
+        while not self.loop.is_closed():
             # Maybe consider having a pool of worker threads, each performing a send in
             # order to increase throughput.
             topic, bytedata = self.event_queue.get()
@@ -273,15 +271,13 @@ class AccessoryDriver(object):
         self.advertiser.register_service(self.mdns_service_info)
 
     def persist(self):
-        """Saves the state of the accessory.
-        """
-        with open(self.persist_file, "w") as fp:
+        """Saves the state of the accessory."""
+        with open(self.persist_file, 'w') as fp:
             self.encoder.persist(fp, self.accessory)
 
     def load(self):
-        """
-        """
-        with open(self.persist_file, "r") as fp:
+        """ """
+        with open(self.persist_file, 'r') as fp:
             self.encoder.load_into(fp, self.accessory)
 
     def pair(self, client_uuid, client_public):
@@ -317,7 +313,7 @@ class AccessoryDriver(object):
         :param client_uuid: The client uuid.
         :type client_uuid: uuid.UUID
         """
-        logger.info("Unpairing client '%s'.", client_uuid)
+        logger.info("Unpairing client %s.", client_uuid)
         self.accessory.remove_paired_client(client_uuid)
         self.persist()
         self.update_advertisement()
@@ -326,7 +322,7 @@ class AccessoryDriver(object):
         """Create an SRP verifier for the accessory's info."""
         # TODO: Move the below hard-coded values somewhere nice.
         ctx = get_srp_context(3072, hashlib.sha512, 16)
-        verifier = SrpServer(ctx, b"Pair-Setup", self.accessory.pincode)
+        verifier = SrpServer(ctx, b'Pair-Setup', self.accessory.pincode)
         self.srp_verifier = verifier
 
     def get_accessories(self):
@@ -381,7 +377,7 @@ class AccessoryDriver(object):
         """
         chars = []
         for id in char_ids:
-            aid, iid = (int(i) for i in id.split("."))
+            aid, iid = (int(i) for i in id.split('.'))
             rep = {HAP_REPR_AID: aid, HAP_REPR_IID: iid}
             char = self.accessory.get_characteristic(aid, iid)
             try:
@@ -463,7 +459,7 @@ class AccessoryDriver(object):
         All of the above are started in separate threads. Accessory thread is set as
         daemon.
         """
-        logger.info("Starting accessory '%s' on address '%s', port '%s'.",
+        logger.info("Starting accessory %s on address %s, port %s.",
                     self.accessory.display_name, self.address, self.port)
 
         # Start sending events to clients. This is done in a daemon thread, because:
@@ -492,21 +488,21 @@ class AccessoryDriver(object):
 
         # Start the accessory so it can do stuff.
         self.accessory.set_sentinel(self.stop_event, self.aio_stop_event,
-                                    self.event_loop)
+                                    self.loop)
         if isinstance(self.accessory, AsyncAccessory):
-            self.accessory_task = self.event_loop.create_task(
+            self.accessory_task = self.loop.create_task(
                 self.accessory.run())
         else:
-            self.accessory_task = self.event_loop.run_in_executor(
+            self.accessory_task = self.loop.run_in_executor(
                 None, self.accessory.run)
 
         logger.info("Starting event loop")
         try:
-            self.event_loop.run_forever()
+            self.loop.run_forever()
         except KeyboardInterrupt:
             pass
         finally:
-            self.event_loop.close()
+            self.loop.close()
             logger.info("Closed event loop.")
 
 
@@ -520,13 +516,13 @@ class AccessoryDriver(object):
         """
         # TODO: This should happen in a different order - mDNS, server, accessory. Need
         # to ensure that sending with a closed server will not crash the app.
-        logger.info("Stoping accessory '%s' on address %s, port %s.",
+        logger.info("Stopping accessory %s on address %s, port %s.",
                     self.accessory.display_name, self.address, self.port)
         logger.debug("Setting stop events, stopping accessory and event sending")
         self.stop_event.set()
-        if not self.event_loop.is_closed():
-            self.event_loop.call_soon_threadsafe(self.aio_stop_event.set)
-            self.event_loop.stop()
+        if not self.loop.is_closed():
+            self.loop.call_soon_threadsafe(self.aio_stop_event.set)
+            self.loop.stop()
         self.accessory.stop()
 
         logger.debug("Stopping mDNS advertising")